requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
Pillow
requests
